@@ -228,7 +228,7 @@ pub fn prio3_client(c: &mut Criterion) {
         bits,
         prio3_input_share_size(&prio3.shard(&measurement).unwrap().1)
     );
-    c.bench_function(&format!("prio3 sum ({} bits)", bits), |b| {
+    c.bench_function(&format!("prio3 sum ({bits} bits)"), |b| {
         b.iter(|| {
             prio3.shard(&measurement).unwrap();
         })
@@ -242,7 +242,7 @@ pub fn prio3_client(c: &mut Criterion) {
         len,
         prio3_input_share_size(&prio3.shard(&measurement).unwrap().1)
     );
-    c.bench_function(&format!("prio3 countvec ({} len)", len), |b| {
+    c.bench_function(&format!("prio3 countvec ({len} len)"), |b| {
         b.iter(|| {
             prio3.shard(&measurement).unwrap();
         })
@@ -257,7 +257,7 @@ pub fn prio3_client(c: &mut Criterion) {
             len,
             prio3_input_share_size(&prio3.shard(&measurement).unwrap().1)
         );
-        c.bench_function(&format!("prio3 parallel countvec ({} len)", len), |b| {
+        c.bench_function(&format!("prio3 parallel countvec ({len} len)"), |b| {
             b.iter(|| {
                 prio3.shard(&measurement).unwrap();
             })
@@ -276,7 +276,7 @@ pub fn prio3_client(c: &mut Criterion) {
             prio3_input_share_size(&prio3.shard(&measurement).unwrap().1)
         );
         c.bench_function(
-            &format!("prio3 fixedpoint16 boundedl2 vec ({} entries)", len),
+            &format!("prio3 fixedpoint16 boundedl2 vec ({len} entries)"),
             |b| {
                 b.iter(|| {
                     prio3.shard(&measurement).unwrap();
@@ -297,10 +297,7 @@ pub fn prio3_client(c: &mut Criterion) {
             prio3_input_share_size(&prio3.shard(&measurement).unwrap().1)
         );
         c.bench_function(
-            &format!(
-                "prio3 fixedpoint16 boundedl2 vec multithreaded ({} entries)",
-                len
-            ),
+            &format!("prio3 fixedpoint16 boundedl2 vec multithreaded ({len} entries)"),
             |b| {
                 b.iter(|| {
                     prio3.shard(&measurement).unwrap();
