@@ -562,8 +562,7 @@ where
 
         if count != 2 {
             return Err(VdafError::Uncategorized(format!(
-                "unexpected message count: got {}; want 2",
-                count,
+                "unexpected message count: got {count}; want 2",
             )));
         }
 
@@ -810,7 +809,7 @@ mod tests {
                 &input.prefix(prefix_len),
                 &[Field128::one(), Field128::one()],
             );
-            assert!(res.is_ok(), "prefix_len={} error: {:?}", prefix_len, res);
+            assert!(res.is_ok(), "prefix_len={prefix_len} error: {res:?}");
         }
 
         // Try evaluating the IDPF keys on incorrect prefixes.
@@ -847,8 +846,7 @@ mod tests {
 
         if expected_output != &output {
             return Err(VdafError::Uncategorized(format!(
-                "eval_idpf(): unexpected output: got {:?}; want {:?}",
-                output, expected_output
+                "eval_idpf(): unexpected output: got {output:?}; want {expected_output:?}"
             )));
         }
 
